separated product endpoints
